% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{boot_se}
\alias{boot_se}
\title{boot_se}
\usage{
boot_se(vec, se = TRUE, b = 1000, n = length(vec[!is.na(vec)]), dig = 2)
}
\arguments{
\item{vec}{}

\item{se}{}

\item{b}{}

\item{n}{}

\item{dig}{}
}
\value{
`out`: booted se or mean of `vec`, depends on supplied `se` param val
}
\description{
boot_se
}
\examples{
boot_se(1:10)
}
