% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{boot_sem}
\alias{boot_sem}
\title{boot_sem
non-parametric bootstrap for standard error of the mean (or just boot mean)}
\usage{
boot_sem(vec, sem = TRUE, b = 1000, n = sum(!is.na(vec)), na.rm = TRUE)
}
\arguments{
\item{vec}{}

\item{b}{number of bootstrap resamples to draw}

\item{n}{number of elements to draw during resampling (defaults to number of non-missing elements in input vector) -- setting `n=1` is jacknife resampling}

\item{se}{logical, if `TRUE` (default), returns standard error of the mean; if `FALSE`, returns the mean of the bootstrap means}
}
\value{
`out`: booted se or mean of `vec`, depends on supplied `se` param val
}
\description{
boot_sem
non-parametric bootstrap for standard error of the mean (or just boot mean)
}
\examples{
boot_se(rnorm(1000))
}
