% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packs.r
\name{check_packs}
\alias{check_packs}
\title{get inventory of attached, installed, loaded, and base packages}
\usage{
check_packs(package_type = "attached")
}
\arguments{
\item{package_type}{can be 'attached', 'installed', 'base', 'loaded', 'all'}
}
\value{
a vector of package names (or a sparse df with a column listing package names for each package type, if `package_type='all'`)
}
\description{
get inventory of attached, installed, loaded, and base packages
}
\examples{
check_packs("attached")   # to see what's directly available
check_packs("installed")   # to see what's in your library 
}
