% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{compare_speed}
\alias{compare_speed}
\title{compare the speed of two expressions}
\usage{
compare_speed(expr1, expr2)
}
\arguments{
\item{expr1}{a quoted piece of code}

\item{expr2}{another quoted piece of code}
}
\value{
length-two vector with names equal to `expr1` and `expr2`, and values equal to the elapsed time after execution of each
}
\description{
NOTE: args must be passed as strings! (for now...)
}
\examples{
compare_speed(expr1="length(1:(10^8))", expr2="length(1:(10^7))")
boosh <- rep(c("bah", "bahbah", "boosh!"), times=1e5)
compare_speed(expr1="sum(grepl('ba', boosh))", 
              expr2="sum(grepl('oo', boosh))") 
}
