% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundled_data.r
\docType{data}
\name{dataset_word_freq}
\alias{dataset_word_freq}
\title{word frequency dataset (5k most frequent english words in COCA)}
\format{a data frame with 5000 rows and five columns
\describe{
  \item{Rank}{a word's frequency rank in COCA}
  \item{Word}{an english word}
  \item{PartOfSpeech}{a part of speech (see COCA documentation for codes)}
  \item{Frequency}{number of occurrences in COCA}
  \item{Dispersion}{dispersion score for the word (see COCA documentation)}
}}
\source{
\url{https://www.wordfrequency.info/}
}
\usage{
dataset_word_freq
}
\description{
a dataset containing corpus freqency, pos, freq rank, and dispersion for the 5k most frequent words in the corpus of contemporary american english (COCA)
}
\keyword{datasets}
