% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{final_seg}
\alias{final_seg}
\title{final segment of string(s) (similar to `base::substr()` but more intuitive)}
\usage{
final_seg(string_vec, nchar, keep_names = FALSE)
}
\arguments{
\item{string_vec}{a vector of character strings}

\item{nchar}{number of characters to return}

\item{keep_names}{for handling named vectors, `FALSE` by default}
}
\value{
final segment of each string in the input vector
}
\description{
get the final segment of length `nchar` for each element of `string_vec`
}
\examples{
final_seg(string_vec=c("hai,","my","name","is","timi!"), nchar=3)
}
