% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.r
\name{get_stop_list}
\alias{get_stop_list}
\title{get one of (currently) five different lists of stopwords}
\usage{
get_stop_list(stop_list = c("una", "bible", "snowball", "onix", "SMART"))
}
\arguments{
\item{stop_list}{can be: 'una', 'bible', 'snowball', 'onix', or 'SMART'}
}
\value{
a vector of common English function words to remove from text before computing on it
}
\description{
get one of (currently) five different lists of stopwords
}
\examples{
get_stop_list(stop_list="una")
}
