% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.r
\name{idf}
\alias{idf}
\title{inverse document frequency: log of num docs / num docs w term}
\usage{
idf(docs, term, lower = TRUE, toss_punct = TRUE, toss_num = TRUE,
  gram_size = 1)
}
\arguments{
\item{docs}{}

\item{term}{}

\item{lower}{}

\item{toss_punct}{}

\item{toss_num}{}

\item{gram_size}{}
}
\value{
num, the inverse doc freq
}
\description{
inverse document frequency: log of num docs / num docs w term
}
\examples{
idf(docs=c("hey im tim", "me too!", "tim tim is tim",) term="tim")
}
