% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_setup.r
\name{lazy_setup}
\alias{lazy_setup}
\title{load my most common set of packages and set plot defaults}
\usage{
lazy_setup(packages = NULL, quiet = TRUE, show_message = TRUE,
  tim = TRUE, set_plot_theme = TRUE, font_size = 14,
  sparse_grid = FALSE, stringsAsFactors = FALSE)
}
\arguments{
\item{packages}{vector of package names, 'defaults' to the five i use most}

\item{quiet}{logical, to suppress package startup messages}

\item{show_message}{logical, for one-line console msg w recap of what happened}

\item{tim}{logical, `TRUE` (the default) if you want to attach `lefftpack::` when calling. **NOTE:** if someone besides me starts using this, i would be happy to change the default to `FALSE` or just remove the argument altogether. just drop me a line.}

\item{set_plot_theme}{logical, if `TRUE` sets to the theme i like}

\item{font_size}{base size for plot theme}

\item{sparse_grid}{logical, if `TRUE` then removes minor grid lines in theme}

\item{stringsAsFactors}{logical, if `FALSE` (default), then session option for `stringsAsFactors` set to `FALSE`}
}
\value{
nothing, used only for side-effects of attaching packages and setting plot theme
}
\description{
('lazy' in the colloquial sense, *not* in the sense of lazy evaluation)
}
\examples{
lefftpack::lazy_setup()
}
