% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lefftpack_data.r
\name{load_lefftpack_data}
\alias{load_lefftpack_data}
\title{load data associated with this package}
\usage{
load_lefftpack_data()
}
\value{
a toy data frame loaded from a csv in the package
}
\description{
load data associated with this package
}
\examples{
boosh <- load_lefftpack_data()
}
