% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{print_lrt_message}
\alias{print_lrt_message}
\title{abbreviated summary of nested model comparison via LRT}
\usage{
print_lrt_message(mob_full = NULL, mob_red = NULL, anova_object = NULL)
}
\arguments{
\item{mob_full}{a model object with k predictors}

\item{mob_red}{a model object with j predictors where j < k}

\item{anova_object}{a comparison of nested models via `anova()`}
}
\value{
mostly for message side-effect, but returns 
the test stat and p-val in a named vector
}
\description{
print just the test stat and p-value for a likelihood ratio test 
assessing relative fit between two nested models (computed w `anova()`)
}
\examples{
fit <- lm(mpg ~ wt + cyl, data=mtcars)
fit_red <- lm(mpg ~ cyl, data=mtcars)

print_lrt_message(mob_full=fit, mob_red=fit_red)

comparison <- anova(fit, fit_red)
print_lrt_message(anova_object=comparison)
}
