% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{pval_classify}
\alias{pval_classify}
\title{pval_classify}
\usage{
pval_classify(x, alpha = 0.05)
}
\arguments{
\item{x}{}

\item{alpha}{}
}
\value{
val
}
\description{
annotate vector of p-values w stars, indicating common signif levels
}
\examples{
boosh 
}
