% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{quiet_attach}
\alias{quiet_attach}
\title{quiet_attach}
\usage{
quiet_attach(packages, shhh = TRUE)
}
\arguments{
\item{packages}{}

\item{shhh}{}
}
\value{
val
}
\description{
quiet_attach
}
\examples{
boosh 
}
