% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{quiet_attach_old}
\alias{quiet_attach_old}
\title{attach packages w/o annoying messages (legacy version)}
\usage{
quiet_attach_old(packages, shhh = TRUE)
}
\arguments{
\item{packages}{}

\item{shhh}{}
}
\value{
val
}
\description{
attach packages w/o annoying messages (legacy version)
}
\examples{
boosh 
}
