% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{r2}
\alias{r2}
\title{round to two digits}
\usage{
r2(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
`x`, rounded to two digits
}
\description{
round to two digits
}
\examples{
r2(c(1.234, 2.34558))
}
