% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.r
\name{remove_stops}
\alias{remove_stops}
\title{remove a list of words from a document (case- and punctuation-sensitive)
(to deal with punctuation etc., call `scrub_doc()` on the `doc` arg before)}
\usage{
remove_stops(doc, stops)
}
\arguments{
\item{doc}{}

\item{stops}{}
}
\value{
the document, with all elements of `stops` removed
}
\description{
remove a list of words from a document (case- and punctuation-sensitive)
(to deal with punctuation etc., call `scrub_doc()` on the `doc` arg before)
}
\examples{
remove_stops(doc="hi i am tim, tim like goat", stops=c("is", "tim"))
}
