% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.r
\name{scrub_doc}
\alias{scrub_doc}
\title{quick func to take a messy doc, clean it up, and return it as a string 
(tokenizes as an intermediate step, but meant for just quick cleaning)}
\usage{
scrub_doc(doc, lower, toss_punct, toss_num)
}
\arguments{
\item{doc}{}

\item{lower}{}

\item{toss_punct}{}

\item{toss_num}{}
}
\value{
the doc as a string, cleaned up
}
\description{
quick func to take a messy doc, clean it up, and return it as a string 
(tokenizes as an intermediate step, but meant for just quick cleaning)
}
\examples{
scrub_doc(doc="H3LL0 me NamE is. t1m timmy", lower=TRUE, toss_punct=TRUE, toss_num=TRUE)
}
