% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.r
\name{text2bigram}
\alias{text2bigram}
\title{convert a document (or doc filepath) to a df of bigrams or bigram counts}
\usage{
text2bigram(doc, is_path = FALSE, stops = NULL, lower = TRUE,
  toss_punct = TRUE, toss_num = TRUE, return_counts = TRUE,
  quiet = FALSE)
}
\arguments{
\item{doc}{}

\item{is_path}{}

\item{stops}{}

\item{lower}{}

\item{toss_punct}{}

\item{toss_num}{}

\item{return_counts}{}

\item{quiet}{}
}
\value{
a data frame with columns for bigram, w1, w2, and count (or just the bigrams and document location indices if `return_counts=FALSE`)
}
\description{
convert a document (or doc filepath) to a df of bigrams or bigram counts
}
\examples{
text2bigram(doc="INDUSTRIAL SOCIETY AND ITS FUTURE  Introduction  1. The Industrial Revolution and its consequences have been a disaster for the human race.")
}
