% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundled_data.r
\docType{data}
\name{text_googmem}
\alias{text_googmem}
\title{the "google memo," written by james damore in 2017.}
\format{a single string containing the entire google memo}
\source{
\url{http://gizmodo.com/exclusive-heres-the-full-10-page-anti-diversity-screed-1797564320}
}
\usage{
text_googmem
}
\description{
the full text of james damore's "google memo," useful for demonstration of text analysis techniques.
}
\keyword{datasets}
