% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundled_data.r
\docType{data}
\name{text_una_manifesto}
\alias{text_una_manifesto}
\title{full text of the unabomber manifesto, written by ted kaczynski ca. 1995}
\format{a length-314 character vector, where each element is a section/paragraph of the unabomber manifesto.}
\source{
\url{http://www.washingtonpost.com/wp-srv/national/longterm/unabomber/manifesto.text.htm}
}
\usage{
text_una_manifesto
}
\description{
full text of the unabomber manifesto, broken down by "sections." originally published by the washington post. useful for practicing text analysis techniques. total word count is ~35k. lol.
}
\keyword{datasets}
