% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.r
\name{tf_idf}
\alias{tf_idf}
\title{tf_idf: how important is a word in a doc, relative to a set of docs (tf*idf)}
\usage{
tf_idf(doc, docs, term, lower = TRUE, toss_punct = TRUE, toss_num = TRUE,
  gram_size = 1)
}
\arguments{
\item{doc}{}

\item{docs}{}

\item{term}{}

\item{lower}{}

\item{toss_punct}{}

\item{toss_num}{}

\item{gram_size}{}
}
\value{
num, term freq times inv doc freq
}
\description{
tf_idf: how important is a word in a doc, relative to a set of docs (tf*idf)
}
\examples{
tf_idf(doc="im tim", docs=c("it is tim", "im tim"), term="tim")
}
