% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.r
\name{word_count_est}
\alias{word_count_est}
\title{estimate the number of words in a document (total or unique)}
\usage{
word_count_est(doc, unique = FALSE)
}
\arguments{
\item{doc}{string or vector of strings or list of strings}

\item{unique}{boolean}
}
\value{
numeric, word count (total or unique)
}
\description{
estimate the number of words in a document (total or unique)
}
\examples{
word_count_est(doc="my name is tim and blah blah blah", unique=TRUE)
}
