% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{bootCI}
\alias{bootCI}
\title{calculate 95% confidence intervals via non-parametric bootstrap}
\usage{
bootCI(vec, b = 1000, n = length(vec[!is.na(vec)]), dig = 6,
  narm = TRUE)
}
\arguments{
\item{vec}{}

\item{b}{}

\item{n}{}

\item{dig}{}

\item{narm}{}
}
\value{
val
}
\description{
returns bootstrap mean and other summary stats
}
\examples{
bootCI(rnorm(100), b=100) 
}
