% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_setup.r
\name{lazy_setup}
\alias{lazy_setup}
\title{load my most common set of packages and set plot defaults}
\usage{
lazy_setup(packages = NULL, quiet = TRUE, show_message = TRUE,
  set_plot_theme = TRUE, font_size = 14, sparse_grid = FALSE)
}
\arguments{
\item{packages}{vector of package names, 'defaults' to the five i use most}

\item{quiet}{logical, to suppress package startup messages}

\item{show_message}{logical, for one-line console msg w recap of what happened}

\item{set_plot_theme}{logical, if `TRUE` sets to the theme i like}

\item{font_size}{base size for plot theme}

\item{sparse_grid}{logical, if `TRUE` then removes minor grid lines in theme}
}
\value{
nothing, used only for side-effects of attaching packages and setting plot theme
}
\description{
('lazy' in the colloquial sense, *not* in the sense of lazy evaluation)
}
\examples{
lefftpack::lazy_setup()
}
