% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{lu}
\alias{lu}
\title{number of unique vals}
\usage{
lu(x)
}
\arguments{
\item{x}{an atomic vector}
}
\value{
number of unique vals, as an integer
}
\description{
(just a wrapper around `length(unique())`, for quick interactive + ad hoc use)
}
\examples{
lu(c(1,2,3,1,2,4))
}
