% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet_attach.r
\name{quiet_attach}
\alias{quiet_attach}
\title{attach packages w/o annoying messages}
\usage{
quiet_attach(..., shhh = "failonly")
}
\arguments{
\item{...}{package names, as characters. can be `...` or a vector}

\item{shhh}{'failonly' (the default) or 'silent' or 'quiet' or 'loud'. throws exception if another value is supplied.}
}
\value{
nothing, used only for side-effects of attaching packages and producing console messages about (un)available packages you tried to attach (if shhh!='silent')
}
\description{
attach packages w/o annoying messages
}
\examples{
quiet_attach("dplyr", "lefftpack", "ballr", shhh="failonly")
}
