% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{zscore}
\alias{zscore}
\title{quick zscore func}
\usage{
zscore(x, na.rm = TRUE)
}
\arguments{
\item{x}{}
}
\value{
val
}
\description{
equivalent to a wrapper around `scale()` w defaults, but returns atomic vec
}
\examples{
boosh 
}
